requirements.txt
openai
pandas
tenacity
pyyaml
faker